Use only single file for all dependencies and versions
